Add files via upload
Basically people are very busy in there daily businesess and emotions in there life .often we forget small things (like looking after the home,switching ofbth lights fans etc ) in life while we focus on the different and higher goals in carrier or relationships Automatic Room Lighting System is a microcontroller based project that automatically turn on or off the lights in a room. Electricity, being one of the most important resources, must be utilized carefully.